Extracted TransferView and AccountsView
