+++ CybridSDK/Components/Transfer/Extensions/TransferView_Loading.swift
@@ -1,0 +1,16 @@
+//
+//  TransferView_Loading.swift
+//  CybridSDK
+//
+//  Created by Erick Sanchez on 06/11/23.
+//
+
+import Foundation
+
+extension TransferView {
+
+    internal func transferView_Loading() {
+        let loadingString = self.localizer.localize(with: UIStrings.loadingText)
+        self.createLoaderScreen(text: loadingString)
+    }
+}
