+++ CybridSDKTestApp/Info.plist
@@ -3,11 +3,11 @@
 <plist version="1.0">
 <dict>
 	<key>CybridClientId</key>
-	<string>CybridClientId</string>
+	<string>PgDKI1v06EowGQjORz0wtE6Pvecvr73COHRt0MPBj90</string>
 	<key>CybridClientSecret</key>
-	<string>CybridClientSecret</string>
+	<string>FnxbaFdU8UZyFgLRHr_y238T-hXIaQGhXsd0hXJkNsc</string>
 	<key>CybridCustomerGUID</key>
-	<string>CybridCustomerGUID</string>
+	<string>16d6aa37d74951c65c4510460cfac71c</string>
 	<key>UIApplicationSceneManifest</key>
 	<dict>
 		<key>UIApplicationSupportsMultipleScenes</key>
