+++ CybridSDKTests/Components Tests/Accounts/View/AccountsViewModelTest.swift
@@ -16,7 +16,6 @@ class AccountsViewModelTest: XCTestCase {
     func createViewModel() -> AccountsViewModel {
 
         return AccountsViewModel(
-            cellProvider: AccountsMockViewProvider(),
             dataProvider: self.dataProvider,
             logger: nil)
     }
@@ -37,7 +36,7 @@ class AccountsViewModelTest: XCTestCase {
 
         // -- Then
         XCTAssertNotNil(viewModel)
-        XCTAssertEqual(viewModel.uiState.value, AccountsViewController.ViewState.LOADING)
+        XCTAssertEqual(viewModel.uiState.value, AccountsView.State.LOADING)
     }
 
     func test_getAccounts_Successfully() async {
@@ -145,110 +144,4 @@ class AccountsViewModelTest: XCTestCase {
         XCTAssertFalse(viewModel.balances.value.isEmpty)
         XCTAssertEqual(viewModel.accountTotalBalance.value, "$40,040,397.82 USD")
     }
-
-    // MARK: TableView Delegation Test
-    func test_TableViewRows() {
-
-        // -- Given
-        let controller = AccountsViewController()
-        let tableView = controller.accountsTable
-        let viewModel = AccountsViewModel(
-                        cellProvider: controller,
-                        dataProvider: self.dataProvider,
-                        logger: nil)
-
-        // -- When
-        viewModel.assets = AssetBankModel.cryptoAssets
-        viewModel.accounts = AccountBankModel.mock
-        viewModel.prices = .mockPrices
-        viewModel.createAccountsFormatted()
-        tableView.reloadData()
-
-        // Then
-        XCTAssertEqual(viewModel.tableView(tableView, numberOfRowsInSection: 0), 3)
-    }
-
-    func test_TableViewHeader() {
-
-        // -- Given
-        let controller = AccountsViewController()
-        let tableView = controller.accountsTable
-        let viewModel = AccountsViewModel(
-                        cellProvider: controller,
-                        dataProvider: self.dataProvider,
-                        logger: nil)
-
-        // -- When
-        viewModel.assets = AssetBankModel.cryptoAssets
-        viewModel.accounts = AccountBankModel.mock
-        viewModel.prices = .mockPrices
-        viewModel.createAccountsFormatted()
-        let headerView = viewModel.tableView(tableView, viewForHeaderInSection: 0)
-
-        // -- Then
-        XCTAssertNotNil(headerView)
-        XCTAssertTrue(headerView!.isKind(of: AccountsHeaderCell.self))
-    }
-
-    func test_TableViewValidCell() {
-
-        // -- Given
-        let controller = AccountsViewController()
-        let tableView = controller.accountsTable
-        let viewModel = AccountsViewModel(
-                        cellProvider: controller,
-                        dataProvider: self.dataProvider,
-                        logger: nil)
-        let indexPath = IndexPath(item: 0, section: 0)
-
-        // -- When
-        viewModel.assets = AssetBankModel.cryptoAssets
-        viewModel.accounts = AccountBankModel.mock
-        viewModel.prices = .mockPrices
-        viewModel.createAccountsFormatted()
-        tableView.reloadData()
-
-        // -- Then
-        XCTAssertTrue(viewModel.tableView(tableView, cellForRowAt: indexPath).isKind(of: AccountsFiatCell.self))
-    }
-
-    func test_TableView_didSelectRowAtIndex() throws {
-
-        // -- Given
-        let controller = AccountsViewController()
-        let tableView = controller.accountsTable
-        let viewModel = AccountsViewModel(
-                        cellProvider: controller,
-                        dataProvider: self.dataProvider,
-                        logger: nil)
-        let indexPath = IndexPath(item: 0, section: 0)
-        let alertExpectation = XCTestExpectation(description: "testAlertShouldAppear")
-
-        // -- When
-        viewModel.assets = AssetBankModel.cryptoAssets
-        viewModel.accounts = AccountBankModel.mock
-        viewModel.prices = .mockPrices
-        viewModel.createAccountsFormatted()
-        tableView.reloadData()
-
-        viewModel.tableView(tableView, didSelectRowAt: indexPath)
-
-        // Then
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0, execute: {
-            XCTAssertNil(controller.presentingViewController)
-            XCTAssertNil(controller.presentedViewController)
-            XCTAssertNil(controller.navigationController)
-            alertExpectation.fulfill()
-        })
-        wait(for: [alertExpectation], timeout: 1.0)
-    }
-}
-
-class AccountsMockViewProvider: AccountsViewProvider {
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath, withData dataModel: BalanceUIModel) -> UITableViewCell {
-        return UITableViewCell()
-    }
-
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath, withBalance balance: BalanceUIModel) {}
 }
